added AppRoutes file and added JSDoc
