+++ src/front_end/src/index.jsx
@@ -1,13 +1,10 @@
+import React from "react"
 import ReactDOM from "react-dom/client"
-import { StrictMode } from "react"
-import App from "./components/App"
+import App from "./App"
 import "./assets/css/global.scss"
-import { BrowserRouter } from "react-router-dom"
-
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </StrictMode>
+const root = ReactDOM.createRoot(document.getElementById("root"))
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 )
