+++ src/front_end/src/pages/registration/registration.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react"
+import { useEffect } from "react"
 import Header from "../../components/header/header"
 import { RegistrationForm } from "../../components/registrationForm/registrationForm"
 import "./registration.scss"
